# control-plane/docs/swagger.yaml
basePath: /api
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: not implemented
        type: string
    type: object
  api.Friend:
    properties:
      in_match:
        example: false
        type: boolean
      online:
        example: true
        type: boolean
      persona_name:
        example: bob
        type: string
      steam_id:
        example: "76561198000000001"
        type: string
    type: object
  api.LoginRequest:
    properties:
      password:
        example: hunter2
        type: string
      username:
        example: alice
        type: string
    type: object
  api.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  api.RegisterRequest:
    properties:
      password:
        example: hunter2
        type: string
      username:
        example: alice
        type: string
    type: object
  api.Session:
    properties:
      id:
        example: 018f8c2e-1d3a-7c1b-9e2a-0a1b2c3d4e5f
        type: string
      match_id:
        example: 7654321098
        type: integer
      state:
        enum:
        - "OFF"
        - STARTING
        - WATCHING
        - STOPPING
        example: WATCHING
        type: string
      target_steam_id:
        example: "76561198000000001"
        type: string
      webrtc_url:
        example: https://dota.example.com/webrtc/live/match
        type: string
    type: object
  api.SessionRequest:
    properties:
      target_steam_id:
        example: "76561198000000001"
        type: string
    type: object
  api.SteamAccount:
    properties:
      created_at:
        example: "2026-06-12T10:00:00Z"
        type: string
      id:
        example: 018f8c2e-1d3a-7c1b-9e2a-0a1b2c3d4e5f
        type: string
      steam_id:
        example: "76561198000000000"
        type: string
      steam_username:
        example: alice_dota
        type: string
    type: object
  api.SteamAccountRequest:
    properties:
      steam_password:
        example: s3cr3t
        type: string
      steam_username:
        example: alice_dota
        type: string
    type: object
  api.SteamGuardRequest:
    properties:
      code:
        example: K4J9X
        type: string
    type: object
info:
  contact: {}
  description: |-
    Self-hosted service to spectate live Dota 2 matches of Steam friends.
    Auth (register/login) is live; other handlers return 501 until their feature steps land.
  title: Dota Spectator Control Plane API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Log in and receive a JWT
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Log out the current session
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: username taken
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /friends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Friend'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List Steam friends with online and in-match status
      tags:
      - friends
  /sessions:
    post:
      consumes:
      - application/json
      parameters:
      - description: target friend
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: worker busy
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start spectating a friend
      tags:
      - sessions
  /sessions/{id}:
    delete:
      parameters:
      - description: session id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Stop spectating
      tags:
      - sessions
    get:
      parameters:
      - description: session id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Session'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get session status (webrtc_url present once WATCHING)
      tags:
      - sessions
  /sessions/{id}/steamguard:
    post:
      consumes:
      - application/json
      parameters:
      - description: session id
        in: path
        name: id
        required: true
        type: string
      - description: guard code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SteamGuardRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Submit a Steam Guard code for a starting session
      tags:
      - sessions
  /steam/accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SteamAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List linked Steam accounts
      tags:
      - steam
    post:
      consumes:
      - application/json
      parameters:
      - description: steam credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.SteamAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.SteamAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Link a Steam account
      tags:
      - steam
  /steam/accounts/{id}:
    delete:
      parameters:
      - description: steam account id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Remove a linked Steam account
      tags:
      - steam
swagger: "2.0"